Fix for zoom

+++ src/main/java/fi/dy/masa/tweakeroo/mixin/MixinGameRenderer.java
@@ -26,11 +26,11 @@ public abstract class MixinGameRenderer
     }
 
     @Inject(method = "getFOVModifier", at = @At("HEAD"), cancellable = true)
-    private void zoom(float partialTicks, boolean useFOVSetting, CallbackInfoReturnable<Float> cir)
+    private void zoom(float partialTicks, boolean useFOVSetting, CallbackInfoReturnable<Double> cir)
     {
         if (FeatureToggle.TWEAK_ZOOM.getBooleanValue() && Hotkeys.ZOOM_ACTIVATE.getKeybind().isKeybindHeld())
         {
-            cir.setReturnValue((float) Configs.Generic.ZOOM_FOV.getDoubleValue());
+            cir.setReturnValue(Configs.Generic.ZOOM_FOV.getDoubleValue());
             cir.cancel();
         }
     }
